Recommit vector.normalize constant fold change Change cvector to take floats instead of doubles
The luau-compile result reads that vector.normalize on a constant
zero vector is equal to [-nan(ind), -nan(ind), -nan(ind)] which made me
think it wouldn't be equal to what the actual vector library would return.
After testing equality on a constant and non constant normalized zero
vector, it seems it does. So it should be fine to let the zero division
happen.